server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.1
numpy==1.26.4
tensorflow==2.17.0
scikit-learn==1.5.2
python-dotenv==1.0.0
aiofiles==23.2.1
soundfile==0.12.1